+++ Package.swift
@@ -6,6 +6,10 @@ import PackageDescription
 let package = Package(
     name: "LazyContainers",
     
+    platforms: [
+        .macOS(.v10_15),
+    ],
+    
     products: [
         .library(
             name: "LazyContainers",